Dev: refactor update_status, 'report type'

+++ genefab3/db/mongo/status.py
@@ -4,33 +4,31 @@ from datetime import datetime
 
 def update_status(collection, logger=None, accession=None, assay_name=None, sample_name=None, status=None, info=None, warning=None, error=None, **kwargs):
     """Update status of dataset (and, optionally, assay/sample) in `collection`, log with logger"""
-    if sample_name is None:
-        if status in {"failed", "dropped"}:
-            run_mongo_transaction(
-                action="delete_many", collection=collection,
-                query=dict(accession=accession),
-            )
-    replacement_query = {
-        "kind": "dataset" if sample_name is None else "sample",
-        "accession": accession, "sample name": sample_name,
-        "warning": warning,
-    }
-    inserted_data = {
-        "status": status, "info": info, "warning": warning,
+    query = {
+        "status": status, "report type": (
+            "dataset status" if sample_name is None else "parser message"
+        ),
+        "accession": accession, "assay name": assay_name,
+        "sample name": sample_name,
+        "info": info, "warning": warning,
         "error": None if (error is None) else type(error).__name__,
-        "report timestamp": int(datetime.now().timestamp()),
-        "assay name": assay_name,
         "args": getattr(error, "args", []), "kwargs": kwargs,
     }
-    run_mongo_transaction(
-        action="replace", collection=collection,
-        query=replacement_query, data=inserted_data,
-    )
+    if (sample_name is None) and (status in {"failed", "dropped"}):
+        run_mongo_transaction(
+            action="delete_many", collection=collection,
+            query=dict(accession=accession),
+        )
+    else:
+        run_mongo_transaction(
+            action="replace", collection=collection, query=query,
+            data={"report timestamp": int(datetime.now().timestamp())},
+        )
     if logger is not None:
         _lookup = dict(failed="error", dropped="error", warning="warning")
         log_kind = _lookup.get(status, "info")
         message = (
             "; ".join([str(_msg) for _msg in (info, warning, error) if _msg]) +
-            " (" + repr(replacement_query) + ")"
+            " (" + repr(query) + ")"
         )
         getattr(logger, log_kind)(message)
